Bugfix for user registration of multiple identities
When registering an identity, the minid server can fail with a 'not
unique' error when used with Globus. This is because the minid code
is marked as 'unique' on the DB schema and isn't saved when Globus
is used. A better solution would be removing the unique constraint,
but saving the code and simply not emailing the user when they use
Globus is an equivalent hack that won't require a db migration.